Refactor potentials
* Make PairPotential and PairParameters special instances of a base
  Potential and Parameters.
* Refactor Tabulator as PotentialTabulator. Simplify to constant width
  tables. PairPotentialTabulator now **always** shifts the potential,
  and it optionally truncates the force (but does modify the potential).
* Create ForceField class to wrap multiple tabulators for use in Simulation.
