+++ relentless/potential/potential.py
@@ -1,4 +1,4 @@
-__all__ = ['PairParameters','PairPotential','Tabulator']
+__all__ = ['Parameters','Potential','PotentialTabulator']
 
 import abc
 import json
@@ -10,106 +10,18 @@ from relentless.core import PairMatrix
 from relentless.core import FixedKeyDict
 from relentless.core import Variable, DependentVariable, IndependentVariable, DesignVariable
 
-class PairParameters(PairMatrix):
-    """Parameters for pairs of types.
-
-    Defines one or more parameters for a set of types. The parameters can be set
-    per-pair, per-type, or shared between all pairs. The per-pair parameters take
-    precedence over the shared parameters. The per-type parameters are not included in
-    the evaluation of pair parameters, but can be used to set per-pair or shared parameters.
-
-    Parameters
-    ----------
-    types : array_like
-        List of types (A type must be a `str`).
-    params : array_like
-        List of parameters (A parameter must be a `str`).
-
-    Raises
-    ------
-    ValueError
-        If params is initialized as empty
-    TypeError
-        If params does not consist of only strings
-
-    Examples
-    --------
-    Create a coefficient matrix with defined types and params::
-
-        m = PairParameters(types=('A','B'), params=('energy','mass'))
-
-    Set coefficient matrix values by accessing parameter directly::
-
-        m['A','A']['energy'] = 2.0
-
-    Assigning to a pair using `update()` overwrites the specified per-pair parameters::
-
-        m['A','A'].update({'mass':2.5})  #does not reset 'energy' value to `None`
-        m['A','A'].update(mass=2.5)      #equivalent statement
-        >>> print(m['A','A'])
-        {'energy':2.0, 'mass':2.5}
-
-    Assigning to a pair using `=` operator overwrites the specified per-pair parameters
-    and resets the other parameters::
-
-        m['A','A'] = {'mass':2.5}  #does reset 'energy' value to `None`
-        >>> print(m['A','A'])
-        {'energy': None, 'mass':2.5}
-
-    Set coefficient matrix values by setting parameters in full::
-
-        m['A','B'] = {'energy':DesignVariable(value=2.0,high=1.5), 'mass':0.5}
-
-    Set coefficient matrix values by iteratively accessing parameters::
-
-        for p in m.params:
-            m['B','B'][p] = 0.1
-
-    Evaluate (retrieve) pair parameters::
-
-        >>> print(m.evaluate(('B','B')))
-        {'energy':0.1, 'mass':0.1}
-
-    Utilizing `evaluate()` computes the defined values of all objects,
-    while directly accessing values returns the objects themselves::
-
-        >>> print(m.evaluate(('A','B')))
-        {'energy':2.0, 'mass':0.5}
-        >>> print(m['A','B'])
-        {'energy':<relentless.core.DesignVariable object at 0x561124456>, 'mass':0.5}
-
-    Assigning to a type sets the specified per-type parameters::
-
-        m['A'].update(energy=1.0, mass=2.0)
-        >>> print(m['A'])
-        {'energy':1.0, 'mass':2.0}
-
-    Assigning to shared sets the specified shared parameters::
-
-        m.shared['energy'] = 0.5
-        >>> print(m.shared)
-        {'energy':0.5, 'mass':None}
-
-    Shared parameters will be used in `evaluate()` if the per-pair parameter is not set::
-
-        >>> m['B','B'] = {'mass': 0.1}
-        >>> m.shared = {'energy': 0.5}
-        >>> print(m['B','B'])
-        {'energy': None, 'mass': 0.1}
-        >>> print(m.shared)
-        {'energy': 0.5, 'mass': None}
-        >>> print(m.evaluate(('B','B'))
-        {'energy':0.5, 'mass':0.1}
-
-    """
+class Parameters:
     def __init__(self, types, params):
-        super().__init__(types)
+        if len(types) == 0:
+            raise ValueError('Cannot initialize with empty types')
+        if not all(isinstance(t, str) for t in types):
+            raise TypeError('All types must be strings')
+        self.types = tuple(types)
 
         if len(params) == 0:
             raise ValueError('params cannot be initialized as empty')
         if not all(isinstance(p, str) for p in params):
             raise TypeError('All parameters must be strings')
-
         self.params = tuple(params)
 
         # shared params
@@ -120,48 +32,16 @@ class PairParameters(PairMatrix):
         for t in self.types:
             self._per_type[t] = FixedKeyDict(keys=self.params)
 
-        # per-pair params
-        self._per_pair = PairMatrix(types)
-        for key in self:
-            self._per_pair[key] = FixedKeyDict(keys=self.params)
-
-    def evaluate(self, pair):
-        """Evaluate pair parameters.
-
-        Returns a dictionary of the parameters and values for the specified pair.
-
-        Parameters
-        ----------
-        pair : tuple
-            Pair for which the parameters are called
-
-        Returns
-        -------
-        params : `dict`
-            The parameters evaluated for the specified pair
-
-        Raises
-        ------
-        TypeError
-            If a parameter is of an unrecognizable type
-        ValueError
-            If a parameter is not set for the specified pair
-
-        Todo
-        ----
-        1. Allow callable parameters.
-        2. Cache output.
-
-        """
+    def evaluate(self, key):
         params = {}
         for p in self.params:
-            # use pair parameter if set, otherwise use shared parameter
-            if self[pair][p] is not None:
-                v = self[pair][p]
+            # use keyed parameter if set, otherwise use shared parameter
+            if self[key][p] is not None:
+                v = self[key][p]
             elif self.shared[p] is not None:
                 v = self.shared[p]
             else:
-                raise ValueError('Parameter {} is not set for ({},{}).'.format(p,pair[0],pair[1]))
+                raise ValueError('Parameter {} is not set for {}.'.format(p,str(key)))
 
             # evaluate the variable
             if isinstance(v, Variable):
@@ -173,7 +53,7 @@ class PairParameters(PairMatrix):
 
             # final check: error if variable is still not set
             if v is None:
-                raise ValueError('Parameter {} is not set for ({},{}).'.format(p,pair[0],pair[1]))
+                raise ValueError('Parameter {} is not set for {}.'.format(p,str(key)))
 
         return params
 
@@ -218,14 +98,9 @@ class PairParameters(PairMatrix):
         return tuple(d)
 
     def __getitem__(self, key):
-        """Get parameters for the (i,j) pair."""
-        if isinstance(key, str):
-            return self._per_type[key]
-        else:
-            return self._per_pair[key]
+        return self._per_type[key]
 
     def __setitem__(self, key, value):
-        """Set parameters for the (i,j) pair."""
         for p in value:
             if p not in self.params:
                 raise KeyError('Only the known parameters can be set in the coefficient matrix.')
@@ -233,253 +108,36 @@ class PairParameters(PairMatrix):
         self[key].update(value)
 
     def __iter__(self):
-        return iter(self._per_pair)
+        return iter(self._per_type)
 
     def __next__(self):
-        return next(self._per_pair)
+        return next(self._per_type)
 
     @property
     def shared(self):
         """:py:class:`FixedKeyDict`: The shared parameters."""
         return self._shared
 
-class PairPotential(abc.ABC):
-    """Generic pair potential evaluator.
+class Potential(abc.ABC):
+    def __init__(self, types, params, container=None):
+        if container is None:
+            container = PotentialParameters
+        self.coeff = container(types,params)
 
-    A PairPotential object is created with coefficients as a PairParameters object.
-    This abstract base class can be extended in order to evaluate custom force,
-    energy, and derivative/gradient functions.
+    @abc.abstractmethod
+    def energy(self, key, r):
+        pass
 
-    Parameters
-    ----------
-    types : array_like
-        List of types (A type must be a `str`).
-    params : array_like
-        List of parameters (A parameter must be a `str`).
+    @abc.abstractmethod
+    def force(self, key, r):
+        pass
 
-    Todo
-    ----
-    1. Inspect _energy() call signature for parameters.
+    @abc.abstractmethod
+    def derivative(self, key, var, r):
+        pass
 
-    """
-    def __init__(self, types, params):
-        # force in standard potential parameters if they are not explicitly set
-        params = list(params)
-        if 'rmin' not in params:
-            params.append('rmin')
-        if 'rmax' not in params:
-            params.append('rmax')
-        if 'shift' not in params:
-            params.append('shift')
-
-        self.coeff = PairParameters(types, params)
-        for p in self.coeff:
-            self.coeff[p]['rmin'] = False
-            self.coeff[p]['rmax'] = False
-            self.coeff[p]['shift'] = False
-
-    def energy(self, pair, r):
-        """Evaluate energy for a (i,j) pair.
-
-        If an `rmin` or `rmax` value is set, then any energy evaluated at an `r`
-        value greater than `rmax` or less than `rmin` is set to the value of energy
-        evaluated or `rmax` or `rmin`, respectively.
-
-        Additionally, if the `shift` parameter is set to be `True`, all energy values
-        are shifted so that the energy is `rmax` is 0.
-
-        Parameters
-        ----------
-        pair : array_like
-            The pair for which to calculate the energy.
-        r : array_like
-            The location(s) at which to evaluate the energy.
-
-        Returns
-        -------
-        scalar or array_like
-            The energy at the specified location(s).
-            The returned quantity will be a scalar if `r` is scalar
-            or a numpy array if `r` is array_like.
-
-        Raises
-        ------
-        ValueError
-            If any value in `r` is negative.
-        ValueError
-            If the potential is shifted without setting `rmax`.
-
-        """
-        params = self.coeff.evaluate(pair)
-        r,u,scalar_r = self._zeros(r)
-        if any(r < 0):
-            raise ValueError('r cannot be negative')
-
-        # evaluate at points below rmax (if set) first, including rmin cutoff (if set)
-        flags = np.ones(r.shape[0], dtype=bool)
-        if params['rmin'] is not False:
-            range_ = r < params['rmin']
-            flags[range_] = False
-            u[range_] = self._energy(params['rmin'], **params)
-        if params['rmax'] is not False:
-            flags[r > params['rmax']] = False
-        u[flags] = self._energy(r[flags], **params)
-
-        # if rmax is set, truncate or shift depending on the mode
-        if params['rmax'] is not False:
-            # with shifting, move the whole potential up
-            # otherwise, set energy to constant for any r beyond rmax
-            if params['shift']:
-                u[r <= params['rmax']] -= self._energy(params['rmax'], **params)
-            else:
-                u[r > params['rmax']] = self._energy(params['rmax'], **params)
-        elif params['shift'] is True:
-            raise ValueError('Cannot shift potential without rmax')
-
-        # coerce u back into shape of the input
-        if scalar_r:
-            u = u.item()
-        return u
-
-    def force(self, pair, r):
-        """Evaluate force for a (i,j) pair.
-
-        The force is only evaluated for `r` values between `rmin` and `rmax`, if set.
-
-        Parameters
-        ----------
-        pair : array_like
-            The pair for which to calculate the force.
-        r : array_like
-            The location(s) at which to evaluate the force.
-
-        Returns
-        -------
-        scalar or array_like
-            The force at the specified location(s).
-            The returned quantity will be a scalar if `r` is scalar
-            or a numpy array if `r` is array_like.
-
-        Raises
-        ------
-        ValueError
-            If any value in `r` is negative.
-
-        """
-        params = self.coeff.evaluate(pair)
-        r,f,scalar_r = self._zeros(r)
-        if any(r < 0):
-            raise ValueError('r cannot be negative')
-
-        # only evaluate at points inside [rmin,rmax], if specified
-        flags = np.ones(r.shape[0], dtype=bool)
-        if params['rmin'] is not False:
-            flags[r < params['rmin']] = False
-        if params['rmax'] is not False:
-            flags[r > params['rmax']] = False
-        f[flags] = self._force(r[flags], **params)
-
-        # coerce f back into shape of the input
-        if scalar_r:
-            f = f.item()
-        return f
-
-    def derivative(self, pair, var, r):
-        """Evaluate derivative for a (i,j) pair with respect to a variable.
-
-        The derivative is only evaluated for `r` values between `rmin` and `rmax`, if set.
-        The derivative can only be evaluted with respect to a :py:class:`Variable`.
-
-        Parameters
-        ----------
-        pair : array_like
-            The pair for which to calculate the derivative.
-        var : :py:class:`Variable`
-            The variable with respect to which the derivative is to be calculated.
-        r : array_like
-            The location(s) at which to calculate the derivative.
-
-        Returns
-        -------
-        scalar or array_like
-            The derivative at the specified location(s).
-            The returned quantity will be a scalar if `r` is scalar
-            or a numpy array if `r` is array_like.
-
-        Raises
-        ------
-        ValueError
-            If any value in `r` is negative.
-        TypeError
-            If the parameter with respect to which to take the derivative
-            is not a :py:class:`Variable`.
-        ValueError
-            If the potential is shifted without setting `rmax`.
-
-        """
-        params = self.coeff.evaluate(pair)
-        r,deriv,scalar_r = self._zeros(r)
-        if any(r < 0):
-            raise ValueError('r cannot be negative')
-        if not isinstance(var, Variable):
-            raise TypeError('Parameter with respect to which to take the derivative must be a Variable.')
-
-        flags = np.ones(r.shape[0], dtype=bool)
-
-        for p in self.coeff.params:
-            # skip shift parameter
-            if p == 'shift':
-                continue
-
-            # try to take chain rule w.r.t. variable first
-            p_obj = self.coeff[pair][p]
-            if isinstance(p_obj, DependentVariable):
-                dp_dvar = p_obj.derivative(var)
-            elif isinstance(p_obj, IndependentVariable) and var is p_obj:
-                dp_dvar = 1.0
-            else:
-                dp_dvar = 0.0
-
-            # skip when dp_dvar is exactly zero, since this does not contribute
-            if dp_dvar == 0.0:
-                continue
-
-            # now take the parameter derivative
-            if p=='rmin':
-                # rmin deriv
-                flags = r < params['rmin']
-                deriv[flags] += -self._force(params['rmin'], **params)*dp_dvar
-            if p=='rmax':
-                # rmax deriv
-                if params['shift']:
-                    flags = r <= params['rmax']
-                    deriv[flags] += self._force(params['rmax'], **params)*dp_dvar
-                else:
-                    flags = r > params['rmax']
-                    deriv[flags] += -self._force(params['rmax'], **params)*dp_dvar
-            else:
-                # regular parameter derivative
-                below = np.zeros(r.shape[0], dtype=bool)
-                if params['rmin'] is not False:
-                    below = r < params['rmin']
-                    deriv[below] += self._derivative(p, params['rmin'], **params)*dp_dvar
-                above = np.zeros(r.shape[0], dtype=bool)
-                if params['rmax'] is not False:
-                    above = r > params['rmax']
-                    deriv[above] += self._derivative(p, params['rmax'], **params)*dp_dvar
-                elif params['shift']:
-                    raise ValueError('Cannot shift without setting rmax.')
-                flags = np.logical_and(~below, ~above)
-                deriv[flags] += self._derivative(p, r[flags], **params)*dp_dvar
-                if params['shift']:
-                    deriv -= self._derivative(p, params['rmax'], **params)*dp_dvar
-
-        # coerce derivative back into shape of the input
-        if scalar_r:
-            deriv = deriv.item()
-        return deriv
-
-    def _zeros(self, r):
+    @classmethod
+    def _zeros(cls, r):
         """Force input to a 1-dimensional array and make matching array of zeros.
 
         Parameters
@@ -508,18 +166,6 @@ class PairPotential(abc.ABC):
             raise TypeError('Expecting 1D array for r')
         return r,np.zeros_like(r),s
 
-    @abc.abstractmethod
-    def _energy(self, r, **params):
-        pass
-
-    @abc.abstractmethod
-    def _force(self, r, **params):
-        pass
-
-    @abc.abstractmethod
-    def _derivative(self, param, r, **params):
-        pass
-
     def save(self, filename):
         """Saves the coefficient matrix to file as JSON data.
 
@@ -537,66 +183,60 @@ class PairPotential(abc.ABC):
     def __next__(self):
         return next(self.coeff)
 
-class Tabulator:
-    """Combines and tabulates multiple potentials together.
+class PotentialTabulator:
+    """Tabulate a potential."""
+    def __init__(self, rmax, num_r, potentials=None):
+        self.rmax = rmax
+        self.num_r = num_r
 
-    Evaluates accumulated energy and force values for multiple potential functions
-    at different r values, allows regularization of the force and shifting of the energy.
+        if potentials is not None:
+            self._potentials = list(potentials)
+        else:
+            self._potentials = []
 
-    Parameters
-    ----------
-    r : array_like
-        The values at which to evaluate energy and force. Must be a 1-D array,
-        with values continously increasing.
-    fmax : float
-        (Optional) The maximum magnitude of the force for regularization at small r values.
-    fcut : float
-        (Optional) The magnitude of the force for truncation at large r values.
-    shift : bool
-        If 'True', shift the potential (according to value of fcut) (defaults to `True`).
+    @property
+    def potentials(self):
+        return self._potentials
 
-    Raises
-    ------
-    ValueError
-        If fmax is set, and it is not positive.
-    ValueError
-        If fcut is set, and it is not positive.
+    @property
+    def rmax(self):
+        return self._rmax
 
-    """
-    def __init__(self, r, fmax=None, fcut=None, shift=True):
-        if fmax is not None and fmax <= 0:
-            raise ValueError('fmax must be positive')
-        if fcut is not None and fcut < 0:
-            raise ValueError('fcut must be positive')
+    @rmax.setter
+    def rmax(self, val):
+        if val is not None and val < 0:
+            raise ValueError('Maximum radius must be positive.')
+        self._rmax = val
+        self._compute_r = True
 
-        self.r = r
-        self.fmax = fmax
-        self.fcut = fcut
-        self.shift = shift
+    @property
+    def num_r(self):
+        return self._num_r
+
+    @num_r.setter
+    def num_r(self, val):
+        if val is not None and (not isinstance(val,int) or val < 2):
+            raise ValueError('Number of points must be at least 2.')
+        self._num_r = val
+        self._compute_r = True
 
     @property
     def r(self):
         """array_like: The values of r at which to evaluate energy and force."""
+        if self._compute_r:
+            if self.rmax is None:
+                raise ValueError('Maximum radius must be set.')
+            if self.num_r is None:
+                raise ValueError('Number of points must be set.')
+            self._r = np.linspace(0,self.rmax,self.num_r,dtype=np.float64)
+            self._compute_r = False
         return self._r
 
-    @r.setter
-    def r(self, points):
-        points = np.array(points)
-        if points.ndim > 1:
-            raise TypeError('r must be a 1-D array')
-        if not np.all(points[1:] > points[:-1]):
-            raise ValueError('r values must be continuously increasing')
-        self._r = points
-
-    def energy(self, pair, potentials):
-        """Evaluates and accumulates energy for all potentials, for the specified pair.
+    def energy(self, key):
+        """Evaluates and accumulates energy for all potentials.
 
         Parameters
         ----------
-        pair : tuple
-            The type pair (i,j) for which to calculate the energy.
-        potentials : array_like
-            All the potential functions for which to calculate and accumulate the energy.
 
         Returns
         -------
@@ -604,23 +244,19 @@ class Tabulator:
             Total energy at each r value.
 
         """
-        u = np.zeros(self.r.shape, dtype=np.float64)
-        for pot in potentials:
+        u = np.zeros_like(self.r)
+        for pot in self.potentials:
             try:
-                u += pot.energy(pair,self.r)
+                u += pot.energy(key,self.r)
             except KeyError:
                 pass
         return u
 
-    def force(self, pair, potentials):
-        """Evaluates and accumulates force for all potentials, for the specified pair.
+    def force(self, key):
+        """Evaluates and accumulates force for all potentials.
 
         Parameters
         ----------
-        pair : tuple
-            The type pair (i,j) for which to calculate the force.
-        potentials : array_like
-            All the potential functions for which to calculate and accumulate the force.
 
         Returns
         -------
@@ -628,90 +264,30 @@ class Tabulator:
             Total force at each r value.
 
         """
-        f = np.zeros(self.r.shape, dtype=np.float64)
-        for pot in potentials:
+        f = np.zeros_like(self.r)
+        for pot in self.potentials:
             try:
-                f += pot.force(pair,self.r)
+                f += pot.force(key,self.r)
             except KeyError:
                 pass
         return f
 
-    def regularize_force(self, u, f, trim=True):
-        """Regularizes and truncates the accumulated energies and forces.
-
-        When shifting is enabled - if fcut is set, then the energies are shifted to be 0 at rcut;
-        otherwise, they are shifted to be 0 at rmax.
+    def derivative(self, key, var):
+        """Evaluates and accumulates derivative for all potentials.
 
         Parameters
         ----------
-        u : array_like
-            Energies, must have the same shape as r.
-        f : array_like
-            Forces, must have the same shape as r.
-        trim : bool
-            Whether to trim off trailing zeros from the regularized force/energy values,
-            defaults to `True`.
 
         Returns
         -------
         array_like
-            A `nx3` array of columns as r, adjusted u, and regularized f, evaluated in n bins.
-        float or `None`
-            The value of r at which trimming trailing zeros is implemented or could be
-            implemented (if trim is False). Returns the last r value if trimming cannot be performed.
-
-        Raises
-        ------
-        IndexError
-            If the energy array is not the same length as the array of r values.
-        IndexError
-            If the force array is not the same length as the array of r values.
-        UserWarning
-            If rmax is too small to cutoff the potential, i.e. if the force at the end
-            of the potential is larger than fcut.
+            Total force at each r value.
 
         """
-        u = np.atleast_1d(u)
-        f = np.atleast_1d(f)
-        if u.shape != self.r.shape:
-            raise IndexError('Potential must have the same length as r.')
-        if f.shape != self.r.shape:
-            raise IndexError('Force must have the same length as r.')
-
-        # find first point from beginning that is within energy tolerance
-        if self.fmax is not None:
-            cut = np.argmax(np.abs(f) <= self.fmax)
-            if cut > 0:
-                u[:cut] = u[cut] - f[cut]*(self.r[:cut] - self.r[cut])
-                f[:cut] = f[cut]
-
-        # find first point from end with sufficient force and cutoff the potential after it
-        if self.fcut is not None:
-            flags = np.abs(np.flip(f)) >= self.fcut
-            cut = len(f)-1 - np.argmax(flags)
-            if cut < len(f)-1:
-                if self.shift:
-                    u -= u[cut]
-                u[(cut+1):] = 0.
-                f[(cut+1):] = 0.
-            else:
-                warnings.warn('Last tabulated force exceeds fcut, rmax may be too small.', UserWarning)
-                if self.shift:
-                    u -= u[-1]
-        elif self.shift:
-            u -= u[-1]
-
-        # trim off trailing zeros
-        r = self.r.copy()
-        flags = np.abs(np.flip(f)) > 0
-        cut = len(f) - np.argmax(flags)
-        if cut < len(f):
-            rcut = r[cut]
-            if trim:
-                r = r[:(cut+1)]
-                u = u[:(cut+1)]
-                f = f[:(cut+1)]
-        else:
-            rcut = r[-1]
-
-        return np.column_stack((r,u,f)), rcut
+        d = np.zeros(self.r)
+        for pot in self.potentials:
+            try:
+                d += pot.derivative(key,var,self.r)
+            except KeyError:
+                pass
+        return d
